% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra.R
\name{query,Spectra-method}
\alias{query,Spectra-method}
\alias{query}
\alias{The}
\alias{`query`}
\alias{function}
\alias{allows}
\alias{to}
\alias{and}
\alias{filter}
\alias{a}
\alias{`Spectra`}
\alias{object}
\alias{using}
\alias{Mass}
\alias{Spec}
\alias{Query}
\alias{Language}
\alias{[MassQL](https://mwang87.github.io/MassQueryLanguage_Documentation/)}
\alias{expression.}
\title{Query a Spectra object using MassQL}
\usage{
\S4method{query}{Spectra}(x, query = character(), ...)
}
\arguments{
\item{x}{The \code{Spectra} object to query.}

\item{query}{\code{character(1)} with the MassQL query.}

\item{...}{currently ignored.}
}
\description{
Query a Spectra object using MassQL
}
\author{
Johannes Rainer
}
